Add files via upload
2. Consider the problem of the previous question and solve the following using dynamic memory
allocation:
a. User will input the brand name and you have to display every product info of that
particular brand.
b. User will input the type of product and you have to calculate the total asset of that
particular type. (qty*price)
c. Calculate the total asset of the grocery shop.